% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/classify.R
\name{classify}
\alias{classify}
\title{Klasifikace DNA sekvence}
\usage{
classify(name, fold, param, csc, cl)
}
\arguments{
\item{name}{jmeno a cesta k FASTA souboru s hledanou sekvenci}

\item{fold}{slozka kde jsou umisteny soubory referencnich dat}

\item{param}{seznam prahovych hodnot, vystup funkce sampl_1 nebo sampl_2}

\item{csc}{pocet nejblizsich fragmentu, zadava uzivatel}

\item{cl}{informace o shlucich, vystup funkce train}
}
\description{
Zpracuje vstupni soubor, najde k nemu nejblizsi shluky, zjistuje pocet zastupcu nadtridy v nejblizsich shlucich, klasifiku podle prahovych hodnot. Potrebuje k cinnosti knihovnu Biostrings.
}
\examples{
classify("ar-ea-015.fa","slozka/",c(0.5,0.4,0.3),100, cl)
classify("cesta/eu-ne-208.fa","slozka",p,30, cl)
}
\keyword{archea,}
\keyword{classification,}
\keyword{eukaryote}
\keyword{prokaryote,}

