% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/make_lib.R
\name{make_lib}
\alias{make_lib}
\title{Vyvoreni referencni sady priznakovych vektoru}
\usage{
make_lib(fold)
}
\arguments{
\item{fold}{slozka kde jsou umisteny FASTA soubory}
}
\description{
Funkce vezme FASTA soubory ze slozky fold, rozdeli je na fragmenty a spocita priznakove vektory. Vysledna data jsou ve slozce classDNAlib. Potrebuje k cinnosti knihovnu Biostrings.
}
\examples{
make_lib("slozka/")
make_lib("slozka")
}
\keyword{FASTA,}
\keyword{fragments,}
\keyword{frequency}
\keyword{oligonucleotide}

