% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/sampl_1.R
\name{sampl_1}
\alias{sampl_1}
\title{Vzorkovani typu 1}
\usage{
sampl_1(fold, samples)
}
\arguments{
\item{fold}{slozka kde jsou umisteny soubory referencnich dat}

\item{samples}{pocet fragmetu z genomu}
}
\description{
Vybere urcity pocet fragmentu z kazdeho genomu, vraci vhodne hodnoty prahu. Vytvari referencni soubory ve slozce classDNAlib_sampl1.
}
\examples{
param <- sampl_1("slozka/", 30)
param <- sampl_1("slozka", 9)
}
\keyword{random,}
\keyword{sampling}

