% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/sampl_2.R
\name{sampl_2}
\alias{sampl_2}
\title{Vzorkovani typu 2}
\usage{
sampl_2(fold, samples)
}
\arguments{
\item{fold}{slozka kde jsou umisteny soubory referencnich dat}

\item{samples}{procento fragmetu z genomu}
}
\description{
Vybere urcity procento fragmentu z kazdeho genomu, vraci vhodne hodnoty prahu. Vytvari referencni soubory ve slozce classDNAlib_sampl2.
}
\examples{
param <- sampl_2("slozka/", 0.3)
param <- sampl_2("slozka", 0.09)
}
\keyword{random,}
\keyword{sampling}

