% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/train.R
\name{train}
\alias{train}
\title{Trenovani na referencni data}
\usage{
train(fold)
}
\arguments{
\item{fold}{slozka kde jsou umisteny soubory referencnich dat}
}
\description{
Provede shlukovani kmeans, vraci informace o shlucich.
}
\examples{
cl <- train("slozka/")
cl <- train("slozka")
}
\keyword{clustering,}
\keyword{kmeans}

